Draft: Remove code duplication Config
- Move ConfigSource classes into lib/misc
- Change nnfw_lib_misc link to private
