+++ runtime/onert/core/include/util/ConfigSource.h
@@ -18,17 +18,19 @@
 #define __ONERT_UTIL_CONFIG_SOURCE_H__
 
 #include <memory>
-
-#include "IConfigSource.h"
+#include <string>
+#include <unordered_map>
 
 namespace onert
 {
 namespace util
 {
 
-void config_source(std::unique_ptr<IConfigSource> &&source);
-void config_source_ext(std::unique_ptr<IConfigSource> &&source);
+using CfgKeyValues = std::unordered_map<std::string, std::string>;
 
+void setConfigKeyValues(const CfgKeyValues &keyValues);
+
+// TODO Move to different place: it is string utility
 bool toBool(const std::string &val);
 int toInt(const std::string &val);
