+++ runtime/service/npud/CMakeLists.txt
@@ -10,8 +10,8 @@ add_executable(npud ${SOURCES})
 set_target_properties(npud PROPERTIES LINKER_LANGUAGE CXX)
 target_include_directories(npud PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_include_directories(npud PUBLIC ${GLIB2.0_INCLUDE_DIRS})
-target_link_libraries(npud PRIVATE ${GLIB2.0_LIBRARIES})
-target_link_libraries(npud PRIVATE ${LIB_PTHREAD})
+target_link_libraries(npud PRIVATE nnfw_lib_misc)
+target_link_libraries(npud PRIVATE ${GLIB2.0_LIBRARIES} ${LIB_PTHREAD})
 
 if(ENVVAR_NPUD_CONFIG)
   target_compile_definitions(npud PRIVATE ENVVAR_FOR_DEFAULT_CONFIG)
