+++ runtime/service/npud/util/ConfigSource.cc
@@ -15,8 +15,10 @@
  */
 
 #include "ConfigSource.h"
-#include "GeneralConfigSource.h"
-#include "EnvConfigSource.h"
+
+#include <misc/IConfigSource.h>
+#include <misc/GeneralConfigSource.h>
+#include <misc/EnvConfigSource.h>
 
 #include <array>
 #include <algorithm>
@@ -28,11 +30,11 @@ namespace npud
 namespace util
 {
 
+using namespace nnfw::misc;
+
 static std::unique_ptr<IConfigSource> _source;
-static std::unique_ptr<IConfigSource> _source_ext;
 
 void config_source(std::unique_ptr<IConfigSource> &&source) { _source = std::move(source); }
-void config_source_ext(std::unique_ptr<IConfigSource> &&source) { _source_ext = std::move(source); }
 
 static IConfigSource *config_source()
 {
@@ -66,14 +68,6 @@ static std::string getConfigOrDefault(const std::string &key)
 
   // Treat empty string and absence of the value to be the same
   auto ret = config_source()->get(key);
-  if (ret.empty())
-  {
-    // if env is not set, search from external
-    if (_source_ext.get())
-    {
-      ret = _source_ext.get()->get(key);
-    }
-  }
   // if not found search from defaults
   if (ret.empty())
   {
